[#4] refactor: update header as moudle
